backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.25
alembic==1.13.2
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9
PyMuPDF==1.24.4
spacy==3.7.4
cryptography==42.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==5.0.0
python-dotenv==1.0.0
email-validator==2.3.0
